downgrade TOTP to 1.1.0 due to strange bug
also keep new icons layout

+++ applications/plugins/totp/scenes/add_new_token/totp_input_text.h
@@ -10,7 +10,7 @@
 
 typedef struct {
     char* user_input;
-    size_t user_input_length;
+    uint8_t user_input_length;
     void* callback_data;
 } InputTextSceneCallbackResult;
